Auto-format Kotlin code style
