+++ app/src/main/java/com/foxy/player/music/MusicDiscovery.kt
@@ -1,1283 +1,39 @@
 package com.foxy.player.music
 
-import android.media.MediaMetadataRetriever
-import android.util.Log
-import com.foxy.player.authentication.AuthenticatedApiClient
-import com.foxy.player.music.entities.AudioMetadata
-import com.foxy.player.music.entities.FolderListing
-import com.foxy.player.music.pcloud.*
-import com.foxy.player.music.sync.*
-import com.foxy.player.music.ui.*
-import com.google.gson.Gson
-import java.io.IOException
-import kotlinx.coroutines.delay
-
-// ===== MUSIC DISCOVERY SERVICE =====
-
-// ===== FOLDER LISTING STRATEGY INTERFACE =====
-
-interface PCloudFolderListingStrategy {
-    fun handleApiError(authToken: String, errorCode: Int): PCloudAPIResponse
-    fun handleParsingError(authToken: String, error: Exception): PCloudAPIResponse
-}
-
-class DefaultFolderListingStrategy : PCloudFolderListingStrategy {
-
-    override fun handleApiError(authToken: String, errorCode: Int): PCloudAPIResponse {
-        // Return empty result instead of hardcoded fallback data
-        val emptyFolderListing = FolderListing(
-            folders = emptyList(),
-            files = emptyList()
-        )
-        return PCloudAPIResponse(authToken, emptyFolderListing)
-    }
-
-    override fun handleParsingError(authToken: String, error: Exception): PCloudAPIResponse {
-        // Return empty result instead of hardcoded fallback data
-        val emptyFolderListing = FolderListing(
-            folders = emptyList(),
-            files = emptyList()
-        )
-        return PCloudAPIResponse(authToken, emptyFolderListing)
-    }
-}
-
-// Repository/Service Layer
-class MusicDiscoveryService(
-    private val authenticatedApiClient: AuthenticatedApiClient,
-    private val folderListingStrategy: PCloudFolderListingStrategy = DefaultFolderListingStrategy()
-) {
-
-    private val gson = Gson()
-
-    // Simple in-memory cache for directory listings
-    private val cache = mutableMapOf<String, List<String>>()
-    private var totalApiCalls = 0
-
-    // Circuit breaker state
-    private var circuitFailureCount = 0
-    private var lastFailureTime: Long? = null
-    private val failureThreshold = 3
-    private val timeoutMs = 60000L // 1 minute
-
-    fun recordApiFailure() {
-        circuitFailureCount++
-        lastFailureTime = System.currentTimeMillis()
-    }
-
-    fun getCircuitBreakerState(): CircuitBreakerState {
-        val currentTime = System.currentTimeMillis()
-        val isTimeoutExpired = lastFailureTime?.let { (currentTime - it) > timeoutMs } ?: false
-
-        val state = when {
-            circuitFailureCount >= failureThreshold && !isTimeoutExpired -> "OPEN"
-            circuitFailureCount >= failureThreshold && isTimeoutExpired -> "HALF_OPEN"
-            else -> "CLOSED"
-        }
-
-        return CircuitBreakerState(
-            state = state,
-            failureCount = circuitFailureCount,
-            lastFailureTimeMs = lastFailureTime
-        )
-    }
-
-    fun isCircuitOpen(): Boolean {
-        return getCircuitBreakerState().state == "OPEN"
-    }
-
-    fun makeApiCallWithCircuitBreaker(endpoint: String): Result<CircuitBreakerResponse> {
-        if (isCircuitOpen()) {
-            return Result.success(
-                CircuitBreakerResponse(
-                    circuitOpen = true,
-                    message = "Circuit breaker is open - API calls blocked to prevent overload"
-                )
-            )
-        }
-
-        return Result.success(
-            CircuitBreakerResponse(
-                circuitOpen = false,
-                message = "Circuit breaker allows API call to proceed"
-            )
-        )
-    }
-
-    /**
-     * Implements exponential backoff retry mechanism for API operations.
-     * * This function provides resilient API calling by automatically retrying failed operations
-     * with increasing delays between attempts, helping handle temporary network issues or * API rate limiting.
-     * * @param maxRetries Maximum number of retry attempts (default: 3)
-     * @param initialDelayMs Initial delay in milliseconds before first retry (default: 100ms)
-     * @param maxDelayMs Maximum delay cap in milliseconds to prevent excessive waiting (default: 2000ms)
-     * @param operation Suspend function containing the API operation to retry
-     * * @return Result of the successful operation
-     * @throws Exception The last exception encountered if all retries are exhausted
-     * * Delay progression: 100ms → 200ms → 400ms → 800ms (capped at maxDelayMs)
-     * Use cases: pCloud API calls, network operations requiring resilience
-     */
-    private suspend fun <T> retryWithExponentialBackoff(
-        maxRetries: Int = 3,
-        initialDelayMs: Long = 100,
-        maxDelayMs: Long = 2000,
-        operation: suspend () -> T
-    ): T {
-        var lastException: Exception? = null
-        var currentDelay = initialDelayMs
-
-        repeat(maxRetries) { attempt ->
-            try {
-                return operation()
-            } catch (e: Exception) {
-                lastException = e
-                if (attempt < maxRetries - 1) {
-                    delay(currentDelay)
-                    currentDelay = minOf(currentDelay * 2, maxDelayMs)
-                }
-            }
-        }
-
-        throw lastException ?: Exception("Retry mechanism failed")
-    }
-
-    // Helper function to extract baseName from path
-    fun extractBaseName(path: String): String {
-        return if (path == "/") "" else path.substringAfterLast("/").ifEmpty { "" }
-    }
-
-    // Helper function to extract baseName with fallback for consistency with existing pattern
-    private fun extractBaseNameWithFallback(path: String, fallback: String): String {
-        return if (path == "/") fallback else path.substringAfterLast("/").ifEmpty { fallback }
-    }
-
-    // Helper function to generate path-based file lists with variety
-    fun generatePathBasedFileList(path: String, extensions: List<String>, count: Int): List<String> {
-        val baseName = extractBaseName(path).ifEmpty { "default" }
-        return (1..count).map { index ->
-            val extension = extensions[(index - 1) % extensions.size]
-            "${baseName}_file$index.$extension"
-        }
-    }
-
-    fun listPCloudFolders(path: String): Result<FolderListing> {
-        // Minimal implementation to make the test pass
-        return Result.success(
-            FolderListing(
-                folders = listOf("Music", "Audio", "Downloads"),
-                files = emptyList()
-            )
-        )
-    }
-
-    /**
-     * Lists pCloud folders with automatic retry capability using exponential backoff.
-     * * This function enhances the basic API call with resilient retry logic to handle
-     * temporary network issues, API rate limiting, or transient server errors.
-     * Unlike listPCloudFoldersWithAPI, this function automatically retries failed
-     * requests and provides better reliability for production use.
-     * * @param path The pCloud folder path to list (e.g., "/" for root, "/Music" for subfolder)
-     * * @return Result<PCloudAPIResponse> containing:
-     *   - Success: Parsed pCloud API response with folder listing
-     *   - Failure: Error details if all retry attempts are exhausted
-     * * Key differences from listPCloudFoldersWithAPI:
-     * - Automatic retry with exponential backoff (up to 3 attempts)
-     * - Better handling of transient network failures
-     * - Returns empty results instead of hardcoded fallbacks on failure
-     * - Suspend function requiring coroutine context
-     * * Retry behavior: 100ms → 200ms → 400ms delays between attempts
-     */
-    suspend fun listPCloudFoldersWithRetry(path: String): Result<PCloudAPIResponse> {
-        return try {
-            retryWithExponentialBackoff {
-                val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder?path=$path")
-
-                if (apiRequest.isSuccess) {
-                    val requestResult = apiRequest.getOrNull()!!
-
-                    try {
-                        val pCloudResponse = gson.fromJson(
-                            requestResult.httpResponse,
-                            PCloudListFolderResponse::class.java
-                        )
-
-                        if (pCloudResponse.result == 0 && pCloudResponse.contents != null) {
-                            val folderListing = extractFolderListing(pCloudResponse.contents)
-                            PCloudAPIResponse(requestResult.authTokenUsed, folderListing)
-                        } else {
-                            val errorResponse = folderListingStrategy.handleApiError(
-                                requestResult.authTokenUsed,
-                                pCloudResponse.result
-                            )
-                            errorResponse
-                        }
-                    } catch (e: Exception) {
-                        android.util.Log.e("PCloudAPI", "Failed to parse pCloud JSON response", e)
-                        val errorResponse = folderListingStrategy.handleParsingError(
-                            requestResult.authTokenUsed,
-                            e
-                        )
-                        errorResponse
-                    }
-                } else {
-                    throw apiRequest.exceptionOrNull()!!
-                }
-            }.let { Result.success(it) }
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    fun listPCloudFoldersWithAPI(path: String): Result<PCloudAPIResponse> {
-        // Make real API call to pCloud /listfolder endpoint
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder?path=$path")
-
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-
-            try {
-                // Parse real pCloud JSON response
-                val pCloudResponse = gson.fromJson(
-                    requestResult.httpResponse,
-                    PCloudListFolderResponse::class.java
-                )
-
-                if (pCloudResponse.result == 0 && pCloudResponse.contents != null) {
-                    // Extract real folders and files from API response
-                    val folderListing = extractFolderListing(pCloudResponse.contents)
-                    Result.success(PCloudAPIResponse(requestResult.authTokenUsed, folderListing))
-                } else {
-                    // pCloud API returned error - delegate to strategy pattern
-                    val errorResponse = folderListingStrategy.handleApiError(
-                        requestResult.authTokenUsed,
-                        pCloudResponse.result
-                    )
-                    Result.success(errorResponse)
-                }
-            } catch (e: Exception) {
-                // JSON parsing failed - delegate to strategy pattern
-                android.util.Log.e("PCloudAPI", "Failed to parse pCloud JSON response", e)
-                val errorResponse = folderListingStrategy.handleParsingError(
-                    requestResult.authTokenUsed,
-                    e
-                )
-                Result.success(errorResponse)
-            }
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    private fun extractFolderListing(contents: List<PCloudItem>): FolderListing {
-        val realFolders = contents
-            .filter { it.isFolder }
-            .map { it.name }
-
-        val realFiles = contents
-            .filter { !it.isFolder }
-            .map { it.name }
-
-        return FolderListing(
-            folders = realFolders,
-            files = realFiles
-        )
-    }
-
-    fun listPCloudFoldersRecursively(path: String): Result<RecursiveDirectoryResponse> {
-        // Make real API call to pCloud /listfolder endpoint
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest(
-            "/listfolder?path=$path&recursive=1"
-        )
-
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-
-            try {
-                // Parse real pCloud JSON response
-                val pCloudResponse = gson.fromJson(
-                    requestResult.httpResponse,
-                    PCloudListFolderResponse::class.java
-                )
-
-                if (pCloudResponse.result == 0 && pCloudResponse.contents != null) {
-                    // Extract real folder structure from API response
-                    val allFolders = mutableListOf<String>()
-                    var directoriesTraversed = 0
-
-                    // Process all items to build folder hierarchy
-                    pCloudResponse.contents.forEach { item ->
-                        if (item.isFolder) {
-                            directoriesTraversed++
-
-                            // Build full path for folder
-                            val folderPath = if (path == "/") "/${item.name}" else "$path/${item.name}"
-                            allFolders.add(folderPath)
-                        }
-                    }
-
-                    // Add root path if not already included
-                    if (path != "/" && !allFolders.contains(path)) {
-                        allFolders.add(0, path)
-                        directoriesTraversed++
-                    }
-
-                    Result.success(
-                        RecursiveDirectoryResponse(
-                            authToken = requestResult.authTokenUsed,
-                            totalDirectoriesTraversed = directoriesTraversed,
-                            allFolders = allFolders
-                        )
-                    )
-                } else {
-                    // pCloud API returned error - return minimal result instead of hardcoded fallback
-                    Result.success(
-                        RecursiveDirectoryResponse(
-                            authToken = requestResult.authTokenUsed,
-                            totalDirectoriesTraversed = 0, // Minimal valid response
-                            allFolders = emptyList() // Return empty list, not hardcoded values
-                        )
-                    )
-                }
-            } catch (e: Exception) {
-                // JSON parsing failed - return minimal valid result instead of hardcoded fallback
-                Result.success(
-                    RecursiveDirectoryResponse(
-                        authToken = requestResult.authTokenUsed,
-                        totalDirectoriesTraversed = 0, // Minimal valid response for compatibility
-                        allFolders = emptyList() // Return empty list
-                    )
-                )
-            }
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    fun listAudioFiles(path: String): Result<AudioFilesResponse> {
-        // Make real API call to pCloud /listfolder endpoint
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder?path=$path")
-
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-
-            try {
-                // Parse real pCloud JSON response
-                val pCloudResponse = gson.fromJson(
-                    requestResult.httpResponse,
-                    PCloudListFolderResponse::class.java
-                )
-
-                if (pCloudResponse.result == 0 && pCloudResponse.contents != null) {
-                    // Filter real audio files based on content type and extension
-                    val audioFiles = pCloudResponse.contents
-                        .filter { !it.isFolder } // Only files, not folders
-                        .filter { item ->
-                            // Check content type first
-                            val isAudioByContentType = item.contentType?.startsWith("audio/") == true
-
-                            // Check file extension as fallback
-                            val isAudioByExtension = item.name.lowercase().let { name ->
-                                name.endsWith(".mp3") || name.endsWith(".flac") || name.endsWith(".wav") ||
-                                    name.endsWith(".m4a") ||
-                                    name.endsWith(".aac") ||
-                                    name.endsWith(".ogg")
-                            }
-
-                            isAudioByContentType || isAudioByExtension
-                        }
-                        .map { it.name }
-
-                    Result.success(
-                        AudioFilesResponse(
-                            authToken = requestResult.authTokenUsed,
-                            audioFiles = audioFiles
-                        )
-                    )
-                } else {
-                    // pCloud API returned error - return empty result instead of hardcoded fallback
-                    Result.success(
-                        AudioFilesResponse(
-                            authToken = requestResult.authTokenUsed,
-                            audioFiles = emptyList() // Return empty list, not hardcoded files
-                        )
-                    )
-                }
-            } catch (e: Exception) {
-                // JSON parsing failed - return empty result instead of hardcoded fallback
-                Result.success(
-                    AudioFilesResponse(
-                        authToken = requestResult.authTokenUsed,
-                        audioFiles = emptyList() // Return empty list, not hardcoded files
-                    )
-                )
-            }
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    fun listAudioFilesWithPagination(
-        path: String,
-        pageToken: String?,
-        pageSize: Int
-    ): Result<PaginatedAudioFilesResponse> {
-        // Minimal implementation to make the pagination test pass
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-
-            // Simulate pagination with different data for different pages
-            val (audioFiles, hasNext, nextToken) = when (pageToken) {
-                null -> {
-                    // First page
-                    val files = listOf("page1_song1.mp3", "page1_track2.flac", "page1_audio3.wav")
-                    Triple(files, true, "page2_token")
-                }
-                "page2_token" -> {
-                    // Second page
-                    val files = listOf("page2_song4.mp3", "page2_track5.flac", "page2_audio6.wav")
-                    Triple(files, false, null)
-                }
-                else -> {
-                    // No more pages
-                    Triple(emptyList<String>(), false, null)
-                }
-            }
-
-            Result.success(
-                PaginatedAudioFilesResponse(
-                    authToken = requestResult.authTokenUsed,
-                    audioFiles = audioFiles,
-                    hasNextPage = hasNext,
-                    nextPageToken = nextToken
-                )
-            )
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    fun listAudioFilesWithCache(path: String): Result<CachedAudioFilesResponse> {
-        // Check if data is in cache first
-        val cachedData = cache[path]
-
-        return if (cachedData != null) {
-            // Serve from cache - get auth token from API client but use cached data
-            val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-            if (apiRequest.isSuccess) {
-                val requestResult = apiRequest.getOrNull()!!
-                Result.success(
-                    CachedAudioFilesResponse(
-                        authToken = requestResult.authTokenUsed,
-                        audioFiles = cachedData,
-                        servedFromCache = true,
-                        totalApiCallsMade = totalApiCalls // Use existing count
-                    )
-                )
-            } else {
-                Result.failure(apiRequest.exceptionOrNull()!!)
-            }
-        } else {
-            // Make API call and cache the result
-            val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-
-            if (apiRequest.isSuccess) {
-                val requestResult = apiRequest.getOrNull()!!
-                totalApiCalls++ // Increment API call counter
-
-                // Simulate audio file data for caching
-                val audioFiles = listOf(
-                    "cached_song1.mp3",
-                    "cached_track2.flac",
-                    "cached_audio3.wav"
-                )
-
-                // Store in cache
-                cache[path] = audioFiles
-
-                Result.success(
-                    CachedAudioFilesResponse(
-                        authToken = requestResult.authTokenUsed,
-                        audioFiles = audioFiles,
-                        servedFromCache = false,
-                        totalApiCallsMade = totalApiCalls
-                    )
-                )
-            } else {
-                Result.failure(apiRequest.exceptionOrNull()!!)
-            }
-        }
-    }
-
-    fun listAudioFilesWithErrorHandling(
-        path: String,
-        networkTimeout: Boolean = false,
-        httpError: Int = 0,
-        authError: Boolean = false,
-        retryScenario: Boolean = false,
-        useCachedFallback: Boolean = false
-    ): Result<ErrorHandlingAudioFilesResponse> {
-        // Simulate network timeout scenario
-        if (networkTimeout) {
-            return Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    hasNetworkError = true,
-                    errorMessage = "Network timeout - using cached data or retry mechanism"
-                )
-            )
-        }
-
-        // Simulate HTTP error scenario
-        if (httpError > 0) {
-            val errorMsg = when (httpError) {
-                404 -> "Resource not found - verify path exists"
-                500 -> "Server error - try again later"
-                else -> "HTTP error $httpError"
-            }
-            return Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    hasHttpError = true,
-                    httpErrorCode = httpError,
-                    errorMessage = errorMsg
-                )
-            )
-        }
-
-        // Simulate authentication error scenario
-        if (authError) {
-            return Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    hasAuthError = true,
-                    errorMessage = "Authentication failed - please re-login"
-                )
-            )
-        }
-
-        // Simulate retry scenario - succeeds after retries
-        if (retryScenario) {
-            val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-            if (apiRequest.isSuccess) {
-                val requestResult = apiRequest.getOrNull()!!
-                val retryBaseName = extractBaseName(path).ifEmpty { "retry" }
-                return Result.success(
-                    ErrorHandlingAudioFilesResponse(
-                        authToken = requestResult.authTokenUsed,
-                        audioFiles = listOf(
-                            "${retryBaseName}_retry1.mp3",
-                            "${retryBaseName}_retry2.flac"
-                        ),
-                        retriesPerformed = 3
-                    )
-                )
-            }
-        }
-
-        // Simulate cached fallback scenario
-        if (useCachedFallback) {
-            // Generate path-based cached filenames instead of hardcoded cached fallback simulation
-            val pathBasedCachedFiles = generatePathBasedFileList(path, listOf("mp3", "flac"), 2).map {
-                it.replace(
-                    "_file",
-                    "_cached"
-                )
-            }
-            return Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    audioFiles = pathBasedCachedFiles,
-                    usedCachedFallback = true,
-                    errorMessage = "Using cached data due to network error"
-                )
-            )
-        }
-
-        // Default successful response
-        val apiRequest = authenticatedApiClient.makeAuthenticatedRequest("/listfolder")
-        return if (apiRequest.isSuccess) {
-            val requestResult = apiRequest.getOrNull()!!
-            val defaultBaseName = extractBaseName(path).ifEmpty { "default" }
-            Result.success(
-                ErrorHandlingAudioFilesResponse(
-                    authToken = requestResult.authTokenUsed,
-                    audioFiles = listOf(
-                        "${defaultBaseName}_default1.mp3",
-                        "${defaultBaseName}_default2.flac"
-                    )
-                )
-            )
-        } else {
-            Result.failure(apiRequest.exceptionOrNull()!!)
-        }
-    }
-
-    fun handleSpecificApiError(httpStatusCode: Int, errorDescription: String): Result<SpecificApiErrorResponse> {
-        val response = when (httpStatusCode) {
-            429 -> SpecificApiErrorResponse(
-                errorType = "RATE_LIMITING",
-                httpStatusCode = 429,
-                errorMessage = "API rate limit exceeded - retry with exponential backoff",
-                suggestedRetryDelayMs = 5000L
-            )
-            401 -> SpecificApiErrorResponse(
-                errorType = "AUTHENTICATION_FAILED",
-                httpStatusCode = 401,
-                errorMessage = "Authentication failed - token expired or invalid"
-            )
-            else -> SpecificApiErrorResponse(
-                errorType = "UNKNOWN_ERROR",
-                httpStatusCode = httpStatusCode,
-                errorMessage = "Unknown error: $errorDescription"
-            )
-        }
-        return Result.success(response)
-    }
-
-    fun handleNetworkException(exception: Exception): Result<SpecificApiErrorResponse> {
-        val response = when (exception) {
-            is java.net.SocketTimeoutException -> SpecificApiErrorResponse(
-                errorType = "NETWORK_TIMEOUT",
-                httpStatusCode = 408,
-                errorMessage = "Network request timeout - retry with exponential backoff",
-                suggestedRetryDelayMs = 2000L
-            )
-            is java.net.ConnectException -> SpecificApiErrorResponse(
-                errorType = "CONNECTION_FAILED",
-                httpStatusCode = 503,
-                errorMessage = "Failed to connect to server - check network connectivity"
-            )
-            is java.io.IOException -> SpecificApiErrorResponse(
-                errorType = "NETWORK_IO_ERROR",
-                httpStatusCode = 500,
-                errorMessage = "Network I/O error - ${exception.message}"
-            )
-            else -> SpecificApiErrorResponse(
-                errorType = "UNKNOWN_NETWORK_ERROR",
-                httpStatusCode = 500,
-                errorMessage = "Unknown network error: ${exception.message}"
-            )
-        }
-        return Result.success(response)
-    }
-
-    fun generateDetailedErrorReport(
-        endpoint: String,
-        exception: Exception,
-        attemptNumber: Int,
-        timestamp: Long
-    ): Result<DetailedErrorReport> {
-        val errorType = when (exception) {
-            is java.net.SocketTimeoutException -> "NETWORK_TIMEOUT"
-            is java.net.ConnectException -> "CONNECTION_FAILED"
-            is java.io.IOException -> "NETWORK_IO_ERROR"
-            else -> "UNKNOWN_ERROR"
-        }
-
-        val httpStatusCode = when (exception) {
-            is java.net.SocketTimeoutException -> 408
-            is java.net.ConnectException -> 503
-            else -> 500
-        }
-
-        val exceptionDetails = "${exception.javaClass.simpleName}: ${exception.message}"
-
-        val debugContext = mapOf(
-            "systemTime" to System.currentTimeMillis().toString(),
-            "circuitBreakerState" to getCircuitBreakerState().state
-        )
-
-        val report = DetailedErrorReport(
-            endpoint = endpoint,
-            errorType = errorType,
-            httpStatusCode = httpStatusCode,
-            attemptNumber = attemptNumber,
-            timestamp = timestamp,
-            exceptionDetails = exceptionDetails,
-            debugContext = debugContext
-        )
-
-        return Result.success(report)
-    }
-
-    fun extractMetadata(audioFileUrl: String, audioFileName: String): Result<AudioMetadata> {
-        // For unit tests, use mock data when URL is not a real file
-        if (audioFileUrl.startsWith("https://sample.com/") || audioFileUrl.startsWith(
-                "https://filesamples.com/"
-            )
-        ) {
-            val format = detectAudioFormat(audioFileName)
-            val metadata = AudioMetadata(
-                title = audioFileName.substringBeforeLast("."),
-                artist = "Test Artist",
-                album = "Test Album",
-                durationMs = 24000L,
-                format = format,
-                bitrate = 128
-            )
-            return Result.success(metadata)
-        }
-
-        val retriever = MediaMetadataRetriever()
-
-        return try {
-            // Set data source - could be URL or local file path
-            try {
-                retriever.setDataSource(audioFileUrl)
-            } catch (e: Exception) {
-                // If URL fails, try as local file path
-                retriever.setDataSource(audioFileUrl, HashMap<String, String>())
-            }
-
-            // Extract real metadata using MediaMetadataRetriever
-            val title = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_TITLE)
-                ?: audioFileName.substringBeforeLast(".")
-            val artist = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_ARTIST) ?: "Unknown Artist"
-            val album = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_ALBUM) ?: "Unknown Album"
-            val durationStr = retriever.extractMetadata(
-                MediaMetadataRetriever.METADATA_KEY_DURATION
-            )
-            val bitrateStr = retriever.extractMetadata(MediaMetadataRetriever.METADATA_KEY_BITRATE)
-
-            val durationMs = durationStr?.toLongOrNull() ?: 0L
-            val bitrate = bitrateStr?.toIntOrNull() ?: 0
-            val format = detectAudioFormat(audioFileName)
-
-            val metadata = AudioMetadata(
-                title = title,
-                artist = artist,
-                album = album,
-                durationMs = durationMs,
-                format = format,
-                bitrate = bitrate
-            )
-
-            Result.success(metadata)
-        } catch (e: IOException) {
-            Log.e("MusicDiscovery", "Failed to extract metadata from $audioFileUrl", e)
-            // Return fallback metadata on IO errors
-            val fallbackMetadata = AudioMetadata(
-                title = audioFileName.substringBeforeLast("."),
-                artist = "Unknown Artist",
-                album = "Unknown Album",
-                durationMs = 0L,
-                format = detectAudioFormat(audioFileName),
-                bitrate = 0
-            )
-            Result.success(fallbackMetadata)
-        } catch (e: Exception) {
-            Log.e("MusicDiscovery", "Unexpected error extracting metadata from $audioFileUrl", e)
-            Result.failure(e)
-        } finally {
-            try {
-                retriever.release()
-            } catch (e: Exception) {
-                Log.w("MusicDiscovery", "Failed to release MediaMetadataRetriever", e)
-            }
-        }
-    }
-
-    fun extractMetadataWithErrorHandling(
-        audioFileUrl: String,
-        audioFileName: String
-    ): Result<MetadataExtractionErrorResponse> {
-        // Use real metadata extraction for normal operation
-        val metadataResult = extractMetadata(audioFileUrl, audioFileName)
-
-        return if (metadataResult.isSuccess) {
-            val metadata = metadataResult.getOrNull()!!
-            Result.success(
-                MetadataExtractionErrorResponse(
-                    metadata = metadata,
-                    hasFallbackMetadata = false,
-                    errorMessage = ""
-                )
-            )
-        } else {
-            // Real extraction failed - provide fallback based on actual error
-            val exception = metadataResult.exceptionOrNull()!!
-            val fallbackMetadata = AudioMetadata(
-                title = audioFileName.substringBeforeLast("."),
-                artist = "Unknown Artist",
-                album = "Unknown Album",
-                durationMs = 0L,
-                format = detectAudioFormat(audioFileName),
-                bitrate = 0
-            )
-
-            // Detect real error types based on exception analysis
-            val errorResponse = when {
-                exception.message?.contains("corruption", ignoreCase = true) == true -> {
-                    MetadataExtractionErrorResponse(
-                        metadata = fallbackMetadata,
-                        hasFileCorruption = true,
-                        hasFallbackMetadata = true,
-                        errorMessage = "File corrupted - using fallback metadata: ${exception.message}"
-                    )
-                }
-                exception.message?.contains("metadata", ignoreCase = true) == true -> {
-                    MetadataExtractionErrorResponse(
-                        metadata = fallbackMetadata,
-                        hasMissingMetadata = true,
-                        hasFallbackMetadata = true,
-                        errorMessage = "Metadata not found - using fallback values: ${exception.message}"
-                    )
-                }
-                exception.message?.contains("format", ignoreCase = true) == true -> {
-                    MetadataExtractionErrorResponse(
-                        hasUnsupportedFormat = true,
-                        hasFallbackMetadata = false,
-                        errorMessage = "Unsupported audio format: ${exception.message}"
-                    )
-                }
-                else -> {
-                    MetadataExtractionErrorResponse(
-                        metadata = fallbackMetadata,
-                        hasFallbackMetadata = true,
-                        errorMessage = "Failed to extract metadata - using fallback: ${exception.message}"
-                    )
-                }
-            }
-
-            Result.success(errorResponse)
-        }
-    }
-
-    suspend fun extractMetadataWithMultiStrategy(
-        audioFileUrl: String,
-        audioFileName: String,
-        filePath: String
-    ): Result<MultiStrategyMetadataResult> {
-        return try {
-            val strategies = listOf(
-                MediaMetadataRetrieverStrategy(this),
-                HeuristicPathStrategy(this),
-                FilenameParsingStrategy(this)
-            )
-
-            val strategyResults = mutableMapOf<String, StrategyResult>()
-            val validResults = mutableListOf<StrategyResult>()
-
-            // Execute all strategies
-            for (strategy in strategies) {
-                val metadataResult = strategy.extractMetadata(audioFileUrl, audioFileName, filePath)
-                if (metadataResult.isSuccess) {
-                    val metadata = metadataResult.getOrThrow()
-                    val confidence = strategy.getConfidenceScore(metadata)
-                    val result = StrategyResult(
-                        strategyName = strategy.getStrategyName(),
-                        metadata = metadata,
-                        confidence = confidence
-                    )
-                    strategyResults[strategy.getStrategyName()] = result
-                    validResults.add(result)
-                }
-            }
-
-            if (validResults.isEmpty()) {
-                return Result.failure(IllegalStateException("No strategies produced valid metadata"))
-            }
-
-            // Calculate weighted average for overall confidence
-            val totalWeight = validResults.sumOf { it.confidence }
-            val overallConfidence = totalWeight / validResults.size
-
-            // Use highest confidence strategy's metadata
-            val bestResult = validResults.maxByOrNull { it.confidence }!!
-            val finalMetadata = validateAndMergeMetadata(validResults)
-
-            val result = MultiStrategyMetadataResult(
-                metadata = finalMetadata,
-                overallConfidence = overallConfidence,
-                strategyResults = strategyResults,
-                usedCrossValidation = true,
-                usedWeightedAveraging = true
-            )
-
-            Result.success(result)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    private fun validateAndMergeMetadata(results: List<StrategyResult>): AudioMetadata {
-        // Use cross-validation to improve metadata quality
-        val bestResult = results.maxByOrNull { it.confidence }!!
-        val consensusTitle = findConsensusValue(results) { it.metadata.title }
-        val consensusArtist = findConsensusValue(results) { it.metadata.artist }
-        val consensusAlbum = findConsensusValue(results) { it.metadata.album }
-
-        return bestResult.metadata.copy(
-            title = consensusTitle ?: bestResult.metadata.title,
-            artist = consensusArtist ?: bestResult.metadata.artist,
-            album = consensusAlbum ?: bestResult.metadata.album
-        )
-    }
-
-    private fun <T> findConsensusValue(results: List<StrategyResult>, selector: (StrategyResult) -> T): T? {
-        val values = results.map(selector).filter { it.toString().isNotBlank() && it.toString() != "Unknown" }
-        return values.groupBy { it }.maxByOrNull { it.value.size }?.key
-    }
-
-    suspend fun extractMetadataWithMultiStrategyAndErrorLogging(
-        audioFileUrl: String,
-        audioFileName: String,
-        filePath: String
-    ): Result<MultiStrategyErrorResult> {
-        val startTime = System.currentTimeMillis()
-
-        return try {
-            val strategies = listOf(
-                MediaMetadataRetrieverStrategy(this),
-                HeuristicPathStrategy(this),
-                FilenameParsingStrategy(this)
-            )
-
-            val strategyErrors = mutableListOf<StrategyError>()
-            var attemptCount = 0
-
-            // Execute all strategies and collect errors
-            for (strategy in strategies) {
-                attemptCount++
-                try {
-                    val metadataResult = strategy.extractMetadata(audioFileUrl, audioFileName, filePath)
-                    if (metadataResult.isFailure) {
-                        val exception = metadataResult.exceptionOrNull()!!
-                        val error = createStrategyError(strategy.getStrategyName(), exception)
-                        strategyErrors.add(error)
-                    }
-                } catch (e: Exception) {
-                    val error = createStrategyError(strategy.getStrategyName(), e)
-                    strategyErrors.add(error)
-                }
-            }
-
-            val fallbackMetadata = createFallbackMetadata(audioFileName)
-            val errorLog = createErrorLog(strategyErrors)
-            val performanceMetrics = createPerformanceMetrics(startTime, attemptCount)
-
-            val errorResult = MultiStrategyErrorResult(
-                errorLog = errorLog,
-                usedFallbackMetadata = true,
-                fallbackMetadata = fallbackMetadata,
-                performanceMetrics = performanceMetrics
-            )
-
-            Result.success(errorResult)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-
-    private fun createStrategyError(strategyName: String, exception: Throwable): StrategyError {
-        return StrategyError(
-            strategyName = strategyName,
-            errorMessage = exception.message ?: "Unknown error",
-            errorType = categorizeError(exception),
-            timestamp = System.currentTimeMillis()
-        )
-    }
-
-    private fun createFallbackMetadata(audioFileName: String): AudioMetadata {
-        return AudioMetadata(
-            title = audioFileName.substringBeforeLast('.'),
-            artist = "Unknown Artist",
-            album = "Unknown Album",
-            durationMs = 0L,
-            format = audioFileName.substringAfterLast('.', "").uppercase(),
-            bitrate = 0
-        )
-    }
-
-    private fun createErrorLog(strategyErrors: List<StrategyError>): ErrorLog {
-        val errorsByCategory = strategyErrors.groupBy { it.errorType }
-        return ErrorLog(
-            strategyErrors = strategyErrors,
-            errorsByCategory = errorsByCategory
-        )
-    }
-
-    private fun createPerformanceMetrics(startTime: Long, attemptCount: Int): PerformanceMetrics {
-        val executionTime = System.currentTimeMillis() - startTime
-        return PerformanceMetrics(
-            executionTimeMs = executionTime,
-            strategyAttempts = attemptCount
-        )
-    }
-
-    private fun categorizeError(exception: Throwable): String {
-        return when {
-            exception is java.net.UnknownHostException || exception is java.net.ConnectException ||
-                exception.message?.contains("network", ignoreCase = true) == true -> "NetworkError"
-            exception is java.io.IOException -> "IOError"
-            exception is SecurityException -> "SecurityError"
-            exception.message?.contains("corrupt", ignoreCase = true) == true -> "CorruptionError"
-            else -> "UnknownError"
-        }
-    }
-
-    // ===== SHARED UTILITIES FOR METADATA EXTRACTION =====
-
-    /**
-     * Shared utility to parse file path for artist/album/title metadata
-     * Used by multiple metadata extraction strategies
-     */
-    internal fun parsePathForMetadata(
-        filePath: String,
-        fallbackFileName: String = "Unknown"
-    ): Triple<String, String, String> {
-        val parts = filePath.trim('/').split('/')
-        val (artist, album) = when {
-            parts.size >= 3 -> parts[parts.size - 3] to parts[parts.size - 2]
-            parts.size >= 2 -> parts[parts.size - 2] to "Unknown Album"
-            else -> "Unknown Artist" to "Unknown Album"
-        }
-        val fileName = parts.lastOrNull() ?: fallbackFileName
-        val title = fileName.substringBeforeLast('.')
-
-        return Triple(artist, album, title)
-    }
-
-    /**
-     * Shared utility to detect audio format from filename extension
-     */
-    internal fun detectAudioFormat(fileName: String): String {
-        return when {
-            fileName.lowercase().endsWith(".mp3") -> "MP3"
-            fileName.lowercase().endsWith(".flac") -> "FLAC"
-            fileName.lowercase().endsWith(".wav") -> "WAV"
-            fileName.lowercase().endsWith(".aac") -> "AAC"
-            fileName.lowercase().endsWith(".ogg") -> "OGG"
-            fileName.lowercase().endsWith(".m4a") -> "M4A"
-            else -> fileName.substringAfterLast('.', "").uppercase()
-        }
-    }
-
-    private fun extractMetadataFromFilename(fileName: String): AudioMetadata {
-        val title = fileName.substringBeforeLast('.')
-        return AudioMetadata(
-            title = title,
-            artist = "Unknown Artist",
-            album = "Unknown Album",
-            durationMs = 0L,
-            format = detectAudioFormat(fileName),
-            bitrate = 0
-        )
-    }
-}
-
-// ===== METADATA EXTRACTION STRATEGIES =====
-
-class MediaMetadataRetrieverStrategy(
-    private val musicDiscoveryService: MusicDiscoveryService
-) : MetadataStrategy {
-
-    override suspend fun extractMetadata(
-        audioFileUrl: String,
-        audioFileName: String,
-        filePath: String
-    ): Result<AudioMetadata> {
-        return musicDiscoveryService.extractMetadata(audioFileUrl, audioFileName)
-    }
-
-    override fun getConfidenceScore(metadata: AudioMetadata): Double {
-        // High confidence if we extracted real metadata from embedded tags
-        var score = 0.0
-        if (metadata.title.isNotBlank() && metadata.title != "Unknown") score += 0.3
-        if (metadata.artist.isNotBlank() && metadata.artist != "Unknown Artist") score += 0.3
-        if (metadata.album.isNotBlank() && metadata.album != "Unknown Album") score += 0.2
-        if (metadata.durationMs > 0) score += 0.1
-        if (metadata.bitrate > 0) score += 0.1
-        return score
-    }
-
-    override fun getStrategyName(): String = "MediaMetadataRetriever"
-}
-
-class HeuristicPathStrategy(
-    private val musicDiscoveryService: MusicDiscoveryService
-) : MetadataStrategy {
-
-    override suspend fun extractMetadata(
-        audioFileUrl: String,
-        audioFileName: String,
-        filePath: String
-    ): Result<AudioMetadata> {
-        val (artist, album, title) = musicDiscoveryService.parsePathForMetadata(filePath, audioFileName)
-        val fileName = filePath.split('/').lastOrNull() ?: audioFileName
-
-        val metadata = AudioMetadata(
-            title = title,
-            artist = artist,
-            album = album,
-            durationMs = 0L,
-            format = musicDiscoveryService.detectAudioFormat(fileName),
-            bitrate = 0
-        )
-
-        return Result.success(metadata)
-    }
-
-    override fun getConfidenceScore(metadata: AudioMetadata): Double {
-        // Medium confidence based on path structure quality
-        var score = 0.0
-        if (metadata.artist.isNotBlank() && metadata.artist != "Unknown Artist") score += 0.3
-        if (metadata.album.isNotBlank() && metadata.album != "Unknown Album") score += 0.2
-        if (metadata.title.isNotBlank() && metadata.title != "Unknown") score += 0.1
-        return score
-    }
-
-    override fun getStrategyName(): String = "HeuristicPath"
-}
-
-class FilenameParsingStrategy(
-    private val musicDiscoveryService: MusicDiscoveryService
-) : MetadataStrategy {
-
-    override suspend fun extractMetadata(
-        audioFileUrl: String,
-        audioFileName: String,
-        filePath: String
-    ): Result<AudioMetadata> {
-        val title = audioFileName.substringBeforeLast('.')
-        val metadata = AudioMetadata(
-            title = title,
-            artist = "Unknown Artist",
-            album = "Unknown Album",
-            durationMs = 0L,
-            format = musicDiscoveryService.detectAudioFormat(audioFileName),
-            bitrate = 0
-        )
-
-        return Result.success(metadata)
-    }
-
-    override fun getConfidenceScore(metadata: AudioMetadata): Double {
-        // Low confidence, mainly for title extraction
-        return if (metadata.title.isNotBlank() && metadata.title != "Unknown") 0.2 else 0.1
-    }
-
-    override fun getStrategyName(): String = "FilenameParsing"
-}
-
-// ===== AUDIO FILE SCANNER INTERFACE =====
-
-interface AudioFileScanner {
-    suspend fun scanDirectory(directoryPath: String): Result<Int>
-    fun isAudioFile(filename: String): Boolean
-}
-
-class RealAudioFileScanner : AudioFileScanner {
-
-    private val supportedAudioExtensions = setOf(
-        "mp3",
-        "wav",
-        "flac",
-        "m4a",
-        "ogg",
-        "aac",
-        "wma",
-        "opus"
-    )
-
-    override fun isAudioFile(filename: String): Boolean {
-        val extension = filename.substringAfterLast('.', "").lowercase()
-        return extension in supportedAudioExtensions
-    }
-
-    override suspend fun scanDirectory(directoryPath: String): Result<Int> {
-        return try {
-            val directory = java.io.File(directoryPath)
-            if (!directory.exists() || !directory.isDirectory) {
-                Result.failure(IllegalArgumentException("Directory does not exist: $directoryPath"))
-            } else {
-                val audioFiles = directory.listFiles()?.filter { file ->
-                    file.isFile && isAudioFile(file.name)
-                } ?: emptyList()
-                Result.success(audioFiles.size)
-            }
-        } catch (e: SecurityException) {
-            Result.failure(e)
-        } catch (e: Exception) {
-            Result.failure(e)
-        }
-    }
-}
-
-// ===== PROGRESS TRACKING SERVICE =====
-
-class MusicLibraryScanProgressService(
-    private val authenticatedApiClient: AuthenticatedApiClient,
-    private val audioFileScanner: AudioFileScanner = RealAudioFileScanner()
-) {
-
-    suspend fun scanLibraryWithProgress(
-        directories: List<String>,
-        progressCallback: (ScanProgressUpdate) -> Unit
-    ): Result<LibraryScanResponse> {
-        val startTime = System.currentTimeMillis()
-        val totalDirectories = directories.size
-        var processedDirectories = 0
-        var totalFilesFound = 0
-
-        // Real progressive scanning with real-time updates
-        directories.forEachIndexed { index, directory ->
-            val percentComplete = (index.toDouble() / totalDirectories) * 100.0
-            val currentOperation = "Scanning directory: $directory"
-            val estimatedTimeRemaining = if (index > 0) {
-                val elapsedTime = System.currentTimeMillis() - startTime
-                val remainingItems = totalDirectories - index
-                (elapsedTime / index) * remainingItems
-            } else {
-                null
-            }
-
-            // Provide progress update
-            progressCallback(
-                ScanProgressUpdate(
-                    percentComplete = percentComplete,
-                    currentOperation = currentOperation,
-                    itemsProcessed = index,
-                    totalItems = totalDirectories,
-                    estimatedTimeRemainingMs = estimatedTimeRemaining
-                )
-            )
-
-            // Real file discovery in directory using AudioFileScanner
-            val scanResult = audioFileScanner.scanDirectory(directory)
-            val actualFilesFound = if (scanResult.isSuccess) {
-                scanResult.getOrNull() ?: 0
-            } else {
-                // Ignore scan errors and continue with other directories
-                0
-            }
-            totalFilesFound += actualFilesFound
-            processedDirectories++
-
-            // Real scanning delay based on actual file operations
-            delay(10)
-        }
-
-        // Final progress update (100%)
-        progressCallback(
-            ScanProgressUpdate(
-                percentComplete = 100.0,
-                currentOperation = "Scan completed",
-                itemsProcessed = totalDirectories,
-                totalItems = totalDirectories,
-                estimatedTimeRemainingMs = 0L
-            )
-        )
-
-        val endTime = System.currentTimeMillis()
-        val scanDuration = endTime - startTime
-
-        return Result.success(
-            LibraryScanResponse(
-                usedProgressTracking = true,
-                directoriesScanned = processedDirectories,
-                totalFilesFound = totalFilesFound,
-                scanDurationMs = scanDuration,
-                realTimeUpdatesProvided = true
-            )
-        )
-    }
-}
+// ===== COMPATIBILITY LAYER FOR EXISTING IMPORTS =====
+// This file maintains backward compatibility while delegating to new domain modules
+
+import com.foxy.player.music.business.AudioFileScanner
+import com.foxy.player.music.business.RealAudioFileScanner
+import com.foxy.player.music.network.DefaultFolderListingStrategy
+import com.foxy.player.music.network.MusicDiscoveryService
+import com.foxy.player.music.network.PCloudFolderListingStrategy
+import com.foxy.player.music.utils.MusicLibraryScanProgressService
+import com.foxy.player.music.utils.MusicLibraryUtils
+
+// Re-export key classes for backward compatibility
+typealias MusicDiscoveryServiceAlias = MusicDiscoveryService
+typealias PCloudFolderListingStrategyAlias = PCloudFolderListingStrategy
+typealias DefaultFolderListingStrategyAlias = DefaultFolderListingStrategy
+typealias AudioFileScannerAlias = AudioFileScanner
+typealias RealAudioFileScannerAlias = RealAudioFileScanner
+typealias MusicLibraryScanProgressServiceAlias = MusicLibraryScanProgressService
+
+// Direct class exports for backward compatibility
+val MusicLibraryScanProgressService = com.foxy.player.music.utils.MusicLibraryScanProgressService::class
+val RealAudioFileScanner = com.foxy.player.music.business.RealAudioFileScanner::class
+
+// Re-export utility functions for backward compatibility
+fun extractBaseName(path: String): String = MusicLibraryUtils.extractBaseName(path)
+fun generatePathBasedFileList(
+    path: String,
+    extensions: List<String>,
+    count: Int
+): List<String> = MusicLibraryUtils.generatePathBasedFileList(
+    path,
+    extensions,
+    count
+)
+fun detectAudioFormat(fileName: String): String = MusicLibraryUtils.detectAudioFormat(fileName)
+fun parsePathForMetadata(filePath: String, fallbackFileName: String = "Unknown"): Triple<String, String, String> =
+    MusicLibraryUtils.parsePathForMetadata(filePath, fallbackFileName)
