+++ app/src/test/java/com/foxy/player/music/MusicDiscoveryApiTest.kt
@@ -3,6 +3,7 @@ package com.foxy.player.music
 import com.foxy.player.authentication.AuthRepository
 import com.foxy.player.authentication.AuthenticatedApiClient
 import com.foxy.player.authentication.UserInfo
+import com.foxy.player.music.network.MusicDiscoveryService
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertNotNull
