+++ app/src/test/java/com/foxy/player/music/MusicLibraryScanTest.kt
@@ -2,7 +2,9 @@ package com.foxy.player.music
 
 import com.foxy.player.authentication.AuthRepository
 import com.foxy.player.authentication.AuthenticatedApiClient
-import com.foxy.player.music.sync.*
+import com.foxy.player.music.business.RealAudioFileScanner
+import com.foxy.player.music.sync.ScanProgressUpdate
+import com.foxy.player.music.utils.MusicLibraryScanProgressService
 import java.io.File
 import java.nio.file.Files
 import kotlinx.coroutines.runBlocking
