+++ app/src/test/java/com/foxy/player/music/RealPCloudAPITest.kt
@@ -2,6 +2,7 @@ package com.foxy.player.music
 
 import com.foxy.player.authentication.AuthRepository
 import com.foxy.player.authentication.AuthenticatedApiClient
+import com.foxy.player.music.network.MusicDiscoveryService
 import java.io.File
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNotNull
