+++ music-player-core/tickets/PLY-132-albumentity-data-class.md
@@ -1,0 +1,70 @@
+# AlbumEntity Data Class
+
+**Objective**
+Add @Entity AlbumEntity data class with id, title, artist fields to MusicModels.kt
+
+**Implementation Notes**
+
+* Follow test-driven development approach
+* Keep all domain code in single file for AI context optimization
+* Use existing codebase patterns and conventions
+
+**Definition of Done**
+
+* Tests written and passing
+* Core functionality implemented
+* Integration validated with existing systems
+* Code review completed
+
+---
+## Android Development Context (Auto-Generated)
+
+**Project**: pCloud Music Player (Android)  
+**Technology Stack**: 
+- Language: Kotlin
+- UI Framework: Jetpack Compose
+- Architecture: MVVM with Repository pattern
+- Audio Engine: Media3/ExoPlayer
+- Networking: Retrofit + OkHttp
+- DI: Hilt/Dagger
+- Build System: Gradle
+
+**Development Domain**: music-player-core  
+**Category**: general  
+
+**Key Android Components Likely Needed**:
+- Activities/Fragments for UI
+- Services for background playback
+- Repository for data access (pCloud API)
+- ViewModels for UI state management
+- Compose UI components
+- MediaSession for media controls
+
+**Testing Strategy**:
+- Unit tests with JUnit 5 + Mockito
+- UI tests with Compose Test
+- Integration tests for API calls
+- Instrumented tests for Android components
+
+**Quality Tools**:
+- ktlint for Kotlin style
+- detekt for code quality
+- Android Lint for security/performance
+- Gradle for build validation
+
+---
+*Generated by Spectrum Development Framework for Android*
+
+---
+## Implementation Completed
+- **Ticket**: PLY-132
+- **PR**: https://github.com/appmancer/FoxyPlayer/pull/42
+- **Domain**: music-player-core
+- **TDD Cycles**: 1 completed
+- **Tests**: 0 passing
+- **Files Changed**: 30
+- **Merged**: 2025-09-02T11:34:42+01:00
+- **Branch**: feature/PLY-132-albumentity-data-class (deleted)
+
+This ticket has been completed and deployed through the Centro development workflow.
+The implementation has been merged to dev branch and deployed to staging environment.
