+++ music-player-core/tickets/PLY-132-tdd-history/cycle-1-20250901-181235.json
@@ -1,0 +1,18 @@
+{
+  "cycle_number": 1,
+  "phase": "completed",
+  "test_description": "should verify AlbumEntity data class structure meets all requirements",
+  "started": "2025-09-01T18:07:15+01:00",
+  "ticket": "PLY-132",
+  "confirmed": true,
+  "confirmed_at": "2025-09-01T17:07:15Z",
+  "green_started": "2025-09-01T17:10:33Z",
+  "green_confirmed": true,
+  "green_confirmed_at": "2025-09-01T17:10:33Z",
+  "refactor_started": "2025-09-01T17:10:49Z",
+  "refactor_confirmed": true,
+  "refactor_confirmed_at": "2025-09-01T17:10:50Z",
+  "commit_started": "2025-09-01T17:12:30Z",
+  "completed_at": "2025-09-01T17:12:35Z",
+  "commit_message": "Verify AlbumEntity data class meets all requirements"
+}
